Makes an Uphold service module and uses that for cleaner separation of concerns
