+++ src/app/blocks/wallet/wallet.component.html
@@ -7,6 +7,6 @@
   
   <a  *ngIf="this.reconnect_button === true" mat-flat-button class="mat-primary" href="{{login_url}}">Reconnect to Uphold</a><br>
 
-  <button *ngIf="this.make_card_button === true" mat-flat-button class="mat-accent" (click)="this.make_card()">Make XRP card</button>
+  <button *ngIf="this.make_card_button === true" mat-flat-button class="mat-accent" (click)="this.makeCard()">Make XRP card</button>
   
 </span>
